Major refactoring
July 26 update
