--- auth/auth-keys.json
@@ -1,4 +1,0 @@
-{
-  "client_id": "",
-  "YT_devkey": ""
-}
