+++ config.json
@@ -1,0 +1,5 @@
+{
+  "client_id": "",
+  "YT_devkey": "",
+  "savePath": "backup/"
+}
